eye_care_scraper/requirements.txt
playwright
python-dotenv
psycopg2-binary
requests
beautifulsoup4
